spanner-tools/spanner_noise_maker/requirements.txt
google-cloud-spanner
Faker
python-json-logger
argparse
